requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
exif==1.6.0
werkzeug==2.3.7
python-dotenv==1.0.0 